+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Added
 
 - #2024 support Expert mode on all Android versions supported by Key Mapper (8.0+).
+- #2025 add report bug button to home screen menu.
 
 ## [4.0.3](https://github.com/sds100/KeyMapper/releases/tag/v4.0.3)